pyproject.toml
[tool.poetry]
name = "yashdb"
version = "0.1.0"
description = ""
authors = ["0xChloe <chloe@catastrophesoftware.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.43.2"
httpx = "^0.28.1"
python-dotenv = "^1.0.1"
loguru = "^0.7.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
